+++ src/Network/Network/Models/AzureFirewall/PSAzureFirewall.cs
@@ -78,7 +78,8 @@ namespace Microsoft.Azure.Commands.Network.Models
         public List<string> Zones { get; set; }
 
         public string AllowActiveFTP { get; set; }
-        public string IdentifyTopFatFlows { get; set; }
+
+        public string IdentifyTopFatFlow { get; set; }
 
         [JsonIgnore]
         public string IpConfigurationsText