+++ src/Network/Network/ChangeLog.md
@@ -19,7 +19,7 @@
 --->
 
 ## Upcoming Release
-* Added `IdentifyTopFatFlows` parameter to `AzureFirewall`
+* Added `IdentifyTopFatFlow` parameter to `AzureFirewall`
     - `New-AzFirewall`
 * Enabled Azure Firewall forced tunneling by default (AzureFirewallManagementSubnet and ManagementPublicIpAddress are required) whenever basic sku firewall is created.
     - `New-AzFirewall`